requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
Pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
geopy==2.3.0
python-dotenv==1.0.0
redis==4.6.0
gunicorn==21.2.0
